# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fraud Risk Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/predictions": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Score Transaction",
        "description": "Score one transaction and store it with its risk assessment.\n\nSend one system's record, plus the other systems' records of the same transaction when\nknown. Records for a TransactionID that is already stored are added to it (409 if that\nsystem's record is already there) and the transaction is re-scored.",
        "operationId": "score_transaction_api_v1_predictions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/predictions/batch": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Score Batch",
        "description": "Score several transactions; each succeeds or fails on its own. Each one counts\nagainst the scoring rate limit.",
        "operationId": "score_batch_api_v1_predictions_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPredictionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/predictions/{prediction_id}": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Prediction",
        "operationId": "get_prediction_api_v1_predictions__prediction_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prediction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ingestion/upload": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Upload Month",
        "description": "Load a month: GL, FA and join map files, plus MA as a file or from its REST API.\n\nReturns at once with a queued batch; poll GET /ingestion/batches/{id} for the outcome.",
        "operationId": "upload_month_api_v1_ingestion_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_month_api_v1_ingestion_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/ingestion/batches": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_api_v1_ingestion_batches_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BatchOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestion/batches/{batch_id}": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Get Batch",
        "operationId": "get_batch_api_v1_ingestion_batches__batch_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchItem": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "status": {
            "type": "string",
            "enum": [
              "scored",
              "error"
            ],
            "title": "Status"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          }
        },
        "type": "object",
        "required": [
          "index",
          "status"
        ],
        "title": "BatchItem"
      },
      "BatchOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "files": {
            "additionalProperties": true,
            "type": "object",
            "title": "Files"
          },
          "records": {
            "additionalProperties": true,
            "type": "object",
            "title": "Records"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "transactions_loaded": {
            "type": "integer",
            "title": "Transactions Loaded"
          },
          "transactions_scored": {
            "type": "integer",
            "title": "Transactions Scored"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "created_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "period",
          "method",
          "status",
          "files",
          "records",
          "summary",
          "transactions_loaded",
          "transactions_scored",
          "error",
          "duration_ms",
          "created_by_id",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "BatchOut"
      },
      "BatchPredictionRequest": {
        "properties": {
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/PredictionRequest"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Transactions"
          }
        },
        "type": "object",
        "required": [
          "transactions"
        ],
        "title": "BatchPredictionRequest"
      },
      "BatchPredictionResult": {
        "properties": {
          "scored": {
            "type": "integer",
            "title": "Scored"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "scored",
          "failed",
          "items"
        ],
        "title": "BatchPredictionResult"
      },
...
      "Body_upload_month_api_v1_ingestion_upload_post": {
        "properties": {
          "period": {
            "type": "string",
            "pattern": "^\\d{4}(0[1-9]|1[0-2])$",
            "title": "Period",
            "examples": [
              "202608"
            ]
          },
          "gl": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Gl",
            "description": "GL report (.xml)"
          },
          "fa": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Fa",
            "description": "FA report (.csv)"
          },
          "join_map": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Join Map",
            "description": "Join map (.txt or .csv)"
          },
          "ma": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma",
            "description": "MA server script (.py) or export (.csv, .json)"
          },
          "ma_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma Url",
            "description": "Pull MA from its REST API instead of a file"
          }
        },
        "type": "object",
        "required": [
          "period",
          "gl",
          "fa",
          "join_map"
        ],
        "title": "Body_upload_month_api_v1_ingestion_upload_post"
      },
...
      "Factor": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "contribution": {
            "type": "number",
            "title": "Contribution"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "features",
          "contribution"
        ],
        "title": "Factor"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_BatchOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[BatchOut]"
      },
...
      "PredictionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "model_score": {
            "type": "integer",
            "title": "Model Score"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "risk_band": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Risk Band"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "exposure_usd": {
            "type": "number",
            "title": "Exposure Usd"
          },
          "break_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Break Types"
          },
          "rule_hits": {
            "items": {
              "$ref": "#/components/schemas/RuleHit"
            },
            "type": "array",
            "title": "Rule Hits"
          },
          "top_factors": {
            "items": {
              "$ref": "#/components/schemas/Factor"
            },
            "type": "array",
            "title": "Top Factors"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "transaction_id",
          "model_version",
          "probability",
          "model_score",
          "risk_score",
          "risk_band",
          "priority",
          "exposure_usd",
          "break_types",
          "rule_hits",
          "top_factors",
          "source",
          "latency_ms",
          "created_at"
        ],
        "title": "PredictionOut"
      },
      "PredictionRequest": {
        "properties": {
          "account_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Account Key",
            "examples": [
              "ACC0001"
            ]
          },
          "transaction_date": {
            "type": "string",
            "format": "date",
            "title": "Transaction Date"
          },
          "amount": {
            "type": "number",
            "maximum": 1000000000000.0,
            "minimum": -1000000000000.0,
            "title": "Amount",
            "examples": [
              1250.0
            ]
          },
          "currency": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Currency",
            "examples": [
              "USD"
            ]
          },
          "country": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Country",
            "examples": [
              "US"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description",
            "examples": [
              "Vendor payment"
            ]
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]{1,64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id",
            "description": "Generated (16 hex characters) when left out.",
            "examples": [
              "9F3A5C7E1B2D4F60"
            ]
          },
          "source_system": {
            "type": "string",
            "enum": [
              "gl",
              "ma",
              "fa"
            ],
            "title": "Source System",
            "default": "gl"
          },
          "counterparts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RecordIn"
            },
            "propertyNames": {
              "enum": [
                "gl",
                "ma",
                "fa"
              ]
            },
            "type": "object",
            "title": "Counterparts",
            "description": "The same transaction as recorded by the other systems, when known."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "account_key",
          "transaction_date",
          "amount",
          "currency",
          "country",
          "description"
        ],
        "title": "PredictionRequest"
      },
      "RecordIn": {
        "properties": {
          "account_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Account Key",
            "examples": [
              "ACC0001"
            ]
          },
          "transaction_date": {
            "type": "string",
            "format": "date",
            "title": "Transaction Date"
          },
          "amount": {
            "type": "number",
            "maximum": 1000000000000.0,
            "minimum": -1000000000000.0,
            "title": "Amount",
            "examples": [
              1250.0
            ]
          },
          "currency": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Currency",
            "examples": [
              "USD"
            ]
          },
          "country": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Country",
            "examples": [
              "US"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description",
            "examples": [
              "Vendor payment"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "account_key",
          "transaction_date",
          "amount",
          "currency",
          "country",
          "description"
        ],
        "title": "RecordIn",
        "description": "One system's record of a transaction. Business rules (ID and account formats, supported\ncurrencies and countries, amount range) are checked by the rule engine and reported as\nrule_violation breaks, not rejected here."
      },
...
      "RuleHit": {
        "properties": {
          "break_type": {
            "type": "string",
            "title": "Break Type"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "break_type",
          "reason"
        ],
        "title": "RuleHit"
      },
...
      "ScoreResult": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "model_score": {
            "type": "integer",
            "title": "Model Score"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "risk_band": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Risk Band"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "exposure_usd": {
            "type": "number",
            "title": "Exposure Usd"
          },
          "break_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Break Types"
          },
          "rule_hits": {
            "items": {
              "$ref": "#/components/schemas/RuleHit"
            },
            "type": "array",
            "title": "Rule Hits"
          },
          "top_factors": {
            "items": {
              "$ref": "#/components/schemas/Factor"
            },
            "type": "array",
            "title": "Top Factors"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created": {
            "type": "boolean",
            "title": "Created",
            "description": "False when the record was added to a known transaction."
          },
          "in_systems": {
            "items": {
              "type": "string",
              "enum": [
                "gl",
                "ma",
                "fa"
              ]
            },
            "type": "array",
            "title": "In Systems"
          }
        },
        "type": "object",
        "required": [
          "id",
          "transaction_id",
          "model_version",
          "probability",
          "model_score",
          "risk_score",
          "risk_band",
          "priority",
          "exposure_usd",
          "break_types",
          "rule_hits",
          "top_factors",
          "source",
          "latency_ms",
          "created_at",
          "created",
          "in_systems"
        ],
        "title": "ScoreResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}